+++ test/tests/system/SystemApplicationTest.java
@@ -67,6 +67,7 @@ public class SystemApplicationTest extends ApplicationTest {
 	public void start(Stage stage) throws Exception {
 		frame = new MainApp();
 		this.stage = stage;
+		frame.init();
 		frame.start(this.stage);
 	}